DWIM: Don't condense to single space at the end of the line
The :startinsert has problems with handling that position on the condensed space at the end of the line, anyway, so special handling had to be implemented, anyway. Better make it useable. Yes, there's a separate d<BS> mapping that completely deletes the trailing whitespace, but for repeating (over multiple columns) it's nice to also have the d<Space> mapping do something sensible.